+++ CHANGES.rst
@@ -1,12 +1,20 @@
 History
 =======
 
-0.0.2
+
+1.0.0
 ------
+
+- Fixed the setting of package settings.
+- Allowed packages to override their setting page.
+
+
+0.0.2
+--------
 
 - Allow links to be added to a controlpanel
 
 0.0.1 - unreleased
-----------------
+--------------------
 
 - Create package with ``pcreate -s kotti kotti_controlpanel``.